allow identity translation for subscriptions (#5934)
This commit solves the following problem:

The MQTT broker sends notification about client subscriptions listing the current subscriptions a client has. E.g.
["topic/a", "topic/b"] means that the client changed its subscriptions and the current subscriptions are those listed.

This also means that when a client unsubscribed, the receiver of this notification can calculate it only when it knows the previous state. E.g. in the list above we may know that the client unsubscribed from "topic/c" if we know the previous list and compare them.

Because of the above, when edgeHub gets a subscription change, it check the current subscriptions of the client and unsubscribes from all not listed in the notification. E.g. if device_c has the current subscriptions:
- C2D messages
- Twin result messages
and it receives the notification ["device_c/twinres"], then it unsubscribes from C2D messages.

This reconciliation is not too have in case of a single device. It becomes more complex when there is a nested scenario and the lower level edgeHub subscribes for every device connected to the lower level and its subscription notification may look like the following: ["device_1/twinres", "device_2/twinres", "device_3/twinres", "device_3/c2d"]

In this case the higher level edgeHub needs to check all the devices with nested connection to be able to calculate the current subscription state. 

The current implementation has two ways of handling subscription changes:
- if it sees that it is not from a lower level edgehub, then it checks only the given device whose subscription changed
- if it is a lower level edgehub, then it checks all nested connection (which is much more work if there are lots of devices)

The two methods above do not work with identity translation, when a device_a may have a notification like ["device_a/twinres", "device_b/twinres", "device_c"/twinres]

Processing this needs the algorithm used for nested connection, but that triggers only when the notification change comes for "$edgeHub". 

This change checks if a normal device/module received notification about a topic of a different device (e.g. module_a has a device_1 listed in the subscription change), and then uses the algorithm developed for nested case - it checks all "indirectly" connected devices to set the subscriptions.

This solution is limited: it works only if there is a single module doing identity translation (but also there can be other modules/devices connecting directly to edgeHub handling their own subscriptions). If there is a module_a translating for device_1 and device_2 and there is a module_b translating for device_3 and device_4, this solution will not work. That is because if module_a has a change as ["device_1/twinres", "device_2/twinres"], then what will happen is that edgeHub checks all the indirectly connected devices (device_1-device_4) and reconciles the subscription list with that device set. Because device_3 and device_4 does not have any subscriptions in the list, edgeHub is going to unsubscribe from anything in the name of those devices.

In order to solve this problem, a change would be needed that tries keeping track which module does identity translation for which devices and this is not easy because this information never shows up directly, only on side channels.